feat(navigation): Added navigation persistor & added segments to screens layer
Due to complex logic of rendering initial screen when the app opens up it was decided to split up the whole "screens" layer into "ui", "config" and "model" segments
