+++ package.json
@@ -32,9 +32,9 @@
     "@hookform/resolvers": "^2.9.10",
     "@react-native-community/checkbox": "^0.5.14",
     "@react-native-community/netinfo": "^9.3.7",
-    "@react-navigation/bottom-tabs": "^6.5.2",
-    "@react-navigation/native": "^6.1.1",
-    "@react-navigation/native-stack": "^6.9.7",
+    "@react-navigation/bottom-tabs": "^6.5.4",
+    "@react-navigation/native": "^6.1.13",
+    "@react-navigation/native-stack": "^6.9.9",
     "@reduxjs/toolkit": "^1.9.1",
     "@tamagui/button": "^1.1.6",
     "@tamagui/core": "^1.1.6",
