+++ src/screens/lib/index.ts
@@ -1,0 +1,2 @@
+export { default as navigationService } from './navigationService';
+export * from './types';
