+++ src/screens/lib/types.ts
@@ -1,0 +1,6 @@
+import { ScreenParams, ScreenRoute } from '../config';
+
+export type InitialRoute = {
+  route: ScreenRoute;
+  params: ScreenParams;
+};
