+++ src/screens/model/hooks/useDefaultRoute.ts
@@ -2,10 +2,10 @@ import { SessionModel } from '@entities/session';
 
 import { ScreenRoute } from '../../config';
 
-function useInitialRoute(): ScreenRoute {
+function useDefaultRoute(): ScreenRoute {
   const hasSession = SessionModel.useHasSession();
 
   return hasSession ? 'Applets' : 'Login';
 }
 
-export default useInitialRoute;
+export default useDefaultRoute;
