+++ src/screens/ui/index.ts
@@ -7,3 +7,5 @@ export { default as SignUpScreen } from './SignUpScreen';
 export { default as AboutScreen } from './AboutScreen';
 export { default as SettingsScreen } from './SettingsScreen';
 export { default as ActivityListScreen } from './ActivityListScreen';
+
+export { default as RootNavigator } from './RootNavigator';
