+++ src/shared/lib/storages/AsyncStorage.ts
@@ -1,18 +1,20 @@
-import Storage from './Storage';
+import { createStorage } from './createStorage';
+
+const storage = createStorage('async-storage');
 
 const AsyncStorage = {
   getItem: (key: string): Promise<string | null> => {
-    const value = Storage.getString(key) ?? null;
+    const value = storage.getString(key) ?? null;
 
     return Promise.resolve(value);
   },
   setItem: (key: string, value: string): Promise<unknown> => {
-    Storage.set(key, value);
+    storage.set(key, value);
 
     return Promise.resolve();
   },
   removeItem: (key: string): Promise<void> => {
-    Storage.delete(key);
+    storage.delete(key);
 
     return Promise.resolve();
   },
