--- src/shared/lib/storages/Storage.ts
@@ -1,5 +1,0 @@
-import { MMKV } from 'react-native-mmkv';
-
-const Storage = new MMKV();
-
-export default Storage;
