+++ src/shared/lib/storages/createStorage.ts
@@ -3,7 +3,7 @@ import { MMKV } from 'react-native-mmkv';
 import { STORE_ENCRYPTION_KEY } from '../constants';
 import { throwError } from '../services';
 
-function createSecureStorage(id: string) {
+export function createSecureStorage(id: string) {
   if (!STORE_ENCRYPTION_KEY) {
     throwError(
       '[createSecureStorage]: STORE_ENCRYPTION_KEY has not been provided',
@@ -13,4 +13,6 @@ function createSecureStorage(id: string) {
   return new MMKV({ id, encryptionKey: STORE_ENCRYPTION_KEY });
 }
 
-export default createSecureStorage;
+export function createStorage(id: string) {
+  return new MMKV({ id });
+}
