+++ src/shared/lib/storages/index.ts
@@ -1,3 +1,2 @@
-export { default as Storage } from './Storage';
 export { default as AsyncStorage } from './AsyncStorage';
-export { default as createSecureStorage } from './createSecureStorage';
+export * from './createStorage';
